test_web/urls.py
from django.urls import include, path
from django.contrib import admin

urlpatterns = [
    # Examples:
    # url(r'^$', 'test_web.views.home', name='home'),
    # url(r'^blog/', include('blog.urls')),

    path('admin/', admin.site.urls),
    path('mqtt/', include('django_mqtt.mosquitto.auth_plugin.urls')),
]
